Move stack generation to the UT #35
- Refactoring - moved stack generation to the UI
- Moved UT to UI testing stuff.
- UT :+1:
- fixes #35
