+++ ModelCompilerUI.UnitTest/packages.config
@@ -2,4 +2,14 @@
 <packages>
   <package id="MSTest.TestAdapter" version="1.3.2" targetFramework="net472" />
   <package id="MSTest.TestFramework" version="1.3.2" targetFramework="net472" />
+  <package id="Newtonsoft.Json" version="13.0.1" targetFramework="net472" />
+  <package id="OPCFoundation.NetStandard.Opc.Ua.Core" version="1.4.367.41" targetFramework="net472" />
+  <package id="OPCFoundation.NetStandard.Opc.Ua.Security.Certificates" version="1.4.367.41" targetFramework="net472" />
+  <package id="Portable.BouncyCastle" version="1.8.10" targetFramework="net472" />
+  <package id="System.Buffers" version="4.5.1" targetFramework="net472" />
+  <package id="System.Formats.Asn1" version="5.0.0" targetFramework="net472" />
+  <package id="System.Memory" version="4.5.4" targetFramework="net472" />
+  <package id="System.Numerics.Vectors" version="4.5.0" targetFramework="net472" />
+  <package id="System.Runtime.CompilerServices.Unsafe" version="4.5.3" targetFramework="net472" />
+  <package id="System.ValueTuple" version="4.5.0" targetFramework="net472" />
 </packages>
