+++ ModelCompilerUI/ModelCompilerUI.csproj
@@ -109,6 +109,7 @@
     </Compile>
     <Compile Include="HeaderUpdateTool.cs" />
     <Compile Include="MeasurementUnits.cs" />
+    <Compile Include="ModelCompilerAPI.cs" />
     <Compile Include="ModelCompilerAPIInternal.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="ToForms\GUIHandling.cs" />
