+++ Opc.Ua.ModelCompiler/Properties/AssemblyInfo.cs
@@ -11,6 +11,13 @@ using System.Runtime.InteropServices;
 [assembly: ComVisible(false)]
 [assembly: Guid("93a1f6d8-7b6e-44b6-8e0b-461cded2cb73")]
 [assembly: InternalsVisibleTo(
+  "OOI.ModelCompilerUI, PublicKey=" +
+    "00240000048000009400000006020000002400005253413100040000010001005b97a0972ff6b13a" +
+    "8a9ff9c09503aea0e5e2fe29cb2275a0c0942182f4c3431814b6bc9a556d9fe0d7e7823439c1ba28" +
+    "521f6318e4c936c4461604ef668e9686c2021571b093e1bfba071b373bc56a07a3afdc120c5313d3" +
+    "9a935cda64b759f857ebb3db483641444a5347e1564f8ba6d4fad2f968d3caf9991a4fa6aa019ebe"
+)]
+[assembly: InternalsVisibleTo(
   "OOI.ModelCompilerUI.UnitTest, PublicKey=" +
     "00240000048000009400000006020000002400005253413100040000010001005b97a0972ff6b13a" +
     "8a9ff9c09503aea0e5e2fe29cb2275a0c0942182f4c3431814b6bc9a556d9fe0d7e7823439c1ba28" +
